feat(settings): Change ColorPreviewIcon function parameter type to Long
Description: Updates the `ColorPreviewIcon` function to accept a `Long` parameter for color, instead of `Int`.

+++ features/settings/src/main/java/com/venom/settings/presentation/components/SettingsIcons.kt
@@ -50,7 +50,7 @@ fun UpdateBadge(hasUpdate: Boolean) {
 }
 
 @Composable
-fun ColorPreviewIcon(color: Int) {
+fun ColorPreviewIcon(color: Long) {
     Surface(
         modifier = Modifier.size(SettingsSpacing.colorPreviewSize),
         shape = CircleShape,
